requirements.txt
langchain==0.0.200
openai==0.27.7
python-dotenv==1.0.0
chardet==5.1.0
sentence_transformers